AI_Services/ocrnerNew/requirements.txt
# Hindi OCR NER Requirements
# Install these packages to run the Hindi OCR and Named Entity Recognition system

# Core ML libraries
torch>=2.0.0
transformers>=4.30.0
torch-audio>=2.0.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.5.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0

# Regex and text processing
regex>=2023.0.0

# JSON and file handling
jsonschema>=4.0.0

# Logging and utilities
tqdm>=4.60.0

# Optional: For enhanced image preprocessing
scikit-image>=0.19.0
matplotlib>=3.5.0

# Optional: For web interface (if needed)
fastapi>=0.68.0
uvicorn>=0.15.0

# Optional: For API testing
requests>=2.25.0